+++ FOX/io/__init__.py
@@ -9,13 +9,15 @@ Various io-related functions implemented in Auto-FOX.
 from .read_kf import read_kf
 from .read_xyz import read_multi_xyz
 from .read_pdb import read_pdb
-from .read_prm import (read_prm, write_prm, rename_atom_types)
+from .read_psf import PSFContainer
+from .read_prm import PRMContainer
 from .hdf5_utils import (create_hdf5, create_xyz_hdf5, to_hdf5, from_hdf5)
 
 __all__ = [
     'read_kf',
     'read_multi_xyz',
     'read_pdb',
-    'read_prm', 'write_prm', 'rename_atom_types',
+    'PSFContainer',
+    'PRMContainer',
     'create_hdf5', 'create_xyz_hdf5', 'to_hdf5', 'from_hdf5',
 ]